refactor: move fields logic to concern
